Iota agent proto and stub (#4)
* First cut of iota-svc

* Add missing files

* Fix make file

* Some fixes

* Add topo file to test cli

* Add vencie to bring up too

* Add node to have make cluster option and make file clean up

* Add venice init and naples init

* Fix venice hostnames

* Remove unwanted file

* Some cleanups

* Fixes after proto changes

* Move server code for ref

* Added init and clean up of testbed

* Added new agent proto
added agent server stub
updated make file

+++ tools/scripts/INSTALL.sh
@@ -43,6 +43,12 @@ then
     cleanupNode
 fi
 
+if [ "$1" == "--clean-only" ]
+then
+    cleanupNode
+    exit 0
+fi
+
 for i in tars/pen* ; do docker load -i  $i; done
 docker run --rm --name pen-install -v /var/log/pensando:/var/log/pensando -v /usr/pensando/bin:/host/usr/pensando/bin -v /usr/lib/systemd/system:/host/usr/lib/systemd/system -v /etc/pensando:/host/etc/pensando pen-install -c /initscript
